requirements.txt
asgiref==3.5.0
dj.choices==0.11.0
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
drf-yasg==1.20.0
pytest==7.1.2
pytest-django==4.5.2
serializers==0.2.4
django-filter==21.1.0